\name{ais}
\title{Australian Institute of Sport Dataset}
\usage{data(ais)}
\alias{ais}
\format{
  A dataframe.
}
\description{
  The first two columns are the gender of the athlete
  and their sport. The remaining 11 columns are various measurements
  made on the athletes.
}
\source{
  Cook and Weisberg (1994).
}
\references{
  Cook and Weisberg (1994). An Introduction to Regression Graphics.
    Wiley, New York.
}
\keyword{datasets}
