\name{bad}
\alias{bad}
\title{Example dataset for bandwidth selection}
\usage{data(bad)}
\format{
Data Frame with x and y variables.
}
\description{
Example dataset from Loader (1999).
}
\references{
  Loader, C. (1999). Bandwidth Selection: Classical or Plug-in?
  Annals of Statistics 27.
}
\keyword{datasets}
