\name{border}
\alias{border}
\title{Cricket Batting Dataset}
\usage{data(border)}
\format{
  A dataframe with day (decimalized); not out indicator and score.
  The not out indicator should be used as a censoring variable.
}
\description{
  Scores in 265 innings for Australian batsman Allan Border.
}
\source{
  Compiled from the Cricinfo archives.
}
\references{CricInfo: The Home of Cricket on the Internet.
  \url{https://www.espncricinfo.com/}
}
\keyword{datasets}
