\name{chemdiab}
\alias{chemdiab}
\title{Chemical Diabetes Dataset}
\usage{data(chemdiab)}
\format{
  Data frame with five numeric measurements and categroical response.
}
\description{
  Numeric variables are \code{rw}, \code{fpg},
  \code{ga}, \code{ina} and \code{sspg}. Classifier \code{cc} is the Diabetic
  type.
}
\source{
  Reaven and Miller (1979).
}
\references{
  Reaven, G. M. and Miller, R. G. (1979).
  An attempt to define the nature of chemical diabetes using a multidimensional
  analysis. Diabetologia 16, 17-24.
}
\keyword{datasets}
