\name{claw54}
\alias{claw54}
\title{Claw Dataset}
\usage{data(claw54)}
\format{
  Numeric vector with length 54.
}
\description{
  A random sample of size 54 from the claw density of Marron and Wand
  (1992), as used in Figure 10.5 of Loader (1999).
}
\source{
  Randomly generated.
}
\references{
  Loader, C. (1999). Local Regression and Likelihood. Springer, New York.
  
  Marron, J. S. and Wand, M. P. (1992). Exact mean integrated squared error.
    Annals of Statistics 20, 712-736.
}
\keyword{datasets}
