\name{cldem}
\alias{cldem}
\title{Example data set for classification}
\usage{data(cldem)}
\format{
Data Frame with x and y variables.
}
\description{
Observations from Figure 8.7 of Loader (1999).
}
\references{
  Loader, C. (1999). Local Regression and Likelihood. Springer, New York.
}
\keyword{datasets}
