\name{co2}
\alias{co2}
\title{Carbon Dioxide Dataset}
\usage{data(co2)}
\format{
  Data frame with \code{year}, \code{month} and \code{co2} variables.
}
\description{
  Monthly time series of carbon dioxide measurements at Mauna Loa,
  Hawaii from 1959 to 1990.
}
\source{
  Boden, Sepanski and Stoss (1992).
}
\references{
  Boden, Sepanski and Stoss (1992).
  Trends '91: A compedium of data on global change - Highlights.
  Carbon Dioxide Information Analysis Center, Oak Ridge National Laboratory.
}
\keyword{datasets}
