\name{ethanol}
\title{Exhaust emissions}
\usage{data(ethanol)}
\alias{ethanol}
\format{
  Data frame with NOx, E and C variables.
}
\description{
  NOx exhaust emissions from a single cylinder engine.  Two predictor
  variables are E (the engine's equivalence ratio) and C (Compression
  ratio). 
}
\source{
  Brinkman (1981). Also studied extensively by Cleveland (1993).
}
\references{
  Brinkman, N. D. (1981). Ethanol fuel - a single-cylinder engine study
  of efficiency and exhaust emissions. SAE transactions 90, 1414-1424.
  
  Cleveland, W. S. (1993). Visualizing data. Hobart Press, Summit, NJ.
}
\keyword{datasets}
