\name{expit}
\alias{expit}
\title{
Inverse logistic link function
}
\usage{
expit(x)
}
\description{
  Computes \eqn{e^x/(1+e^x)}{\exp(x)/(1+\exp(x))}. 
  This is the inverse of the logistic link function,
  \eqn{\log(p/(1-p))}.
}
\arguments{
  \item{x}{numeric vector}
}

\keyword{math}
% Converted by Sd2Rd version 0.2-a5.
