\name{formula.locfit}
\alias{formula.locfit}
\title{
Formula from a Locfit object.
}
\usage{
formula(object)
}
\description{Extract the model formula from a locfit object.}
\arguments{
\item{object}{ \code{"locfit"} object.}
}
\value{
Returns the formula from the locfit object.
}
\seealso{
\code{locfit}
}
\keyword{models}
% Converted by Sd2Rd version 0.2-a5.
