\name{gam.slist}
\alias{gam.slist}
\title{Vector of GAM special terms}
\format{
  Character vector.
}
\description{
  This vector adds \code{"lf"} to the default vector of special
  terms recognized by a \code{gam()} model formula.
  To ensure this is recognized, attach the Locfit library with
  \code{library(locfit,first=T)}.
}
\seealso{
  \code{\link{lf}},
  \code{gam}
}
\keyword{datasets}
