\name{geyser}
\alias{geyser}
\title{Old Faithful Geyser Dataset}
\usage{data(geyser)}
\format{
  A numeric vector of length 107.
}
\description{
  The durations of 107 eruptions of the Old Faithful Geyser.
}
\source{
  Scott (1992). Note that several different Old Faithful Geyser datasets
  (including the faithful dataset in R's base library) have been used
  in various places in the statistics literature. The version provided
  here has been used in density estimation and bandwidth selection work.
}
\references{
  Scott, D. W. (1992). Multivariate Density Estimation: Theory,
  Practice and Visualization. Wiley.
}
\keyword{datasets}
