\name{geyser.round}
\alias{geyser.round}
\title{Discrete Old Faithful Geyser Dataset}
\usage{data(geyser.round)}
\format{
  Data Frame with variables \code{duration} and \code{count}.
}
\description{
  This is a variant of the \code{\link{geyser}} dataset, where
  each observation is rounded to the nearest 0.05 minutes, and the
  counts tallied.
}
\source{
  Scott (1992). Note that several different Old Faithful Geyser datasets
  (including the faithful dataset in R's base library) have been used
  in various places in the statistics literature. The version provided
  here has been used in density estimation and bandwidth selection work.
}
\references{
  Scott, D. W. (1992). Multivariate Density Estimation: Theory,
  Practice and Visualization. Wiley.
}
\keyword{datasets}
