\name{heart}
\title{Survival Times of Heart Transplant Recipients}
\usage{data(heart)}
\alias{heart}
\format{
  Data frame with surv, cens and age variables.
}
\description{
  The survival times of 184 participants in the Stanford heart
  transplant program.
}
\source{
  Miller and Halperin (1982).
  The original dataset includes information on additional patients who
  never received a transplant. Other authors reported earlier versions
  of the data.
}
\references{
  Miller, R. G. and Halperin, J. (1992). Regression with censored data.
  Biometrika 69, 521-531.
}
\keyword{datasets}
