\name{insect}
\alias{insect}
\title{Insect Dataset}
\usage{data(insect)}
\format{
  Data frame with \code{lconc} (dosage), \code{deaths}
  (number of deaths) and \code{nins}
  (number of insects) variables.
}
\description{
  An experiment measuring death rates for insects, with 30 insects at
  each of five treatment levels.
}
\source{
  Bliss (1935).
}
\references{
  Bliss (1935).
  The calculation of the dosage-mortality curve.
  Annals of Applied Biology 22, 134-167.
}
\keyword{datasets}
