\name{iris}
\title{Fisher's Iris Data (subset)}
\usage{data(iris)}
\alias{iris}
\format{
  Data frame with species, petal.wid, petal.len, sepal.wid, sepal.len.
}
\description{
  Four measurements on each of fifty flowers of two species of iris
  (Versicolor and Virginica) -- A classification dataset. Fisher's
  original dataset contained a third species (Setosa) which is trivially
  seperable.
}
\source{
  Fisher (1936). Reproduced in Andrews and Herzberg (1985) Chapter 1.

}
\references{
  Andrews, D. F. and Herzberg, A. M. (1985). Data. Springer-Verlag.

  Fisher, R. A. (1936). The Use of Multiple Measurements in Taxonomic
  Problems. Annals of Eugenics 7, Part II. 179-188.
}
\keyword{datasets}
