\name{kangaroo}
\alias{kangaroo}
\title{Kangaroo skull measurements dataset}
\usage{data(kangaroo)}
\format{
  Data frame with measurements on the skulls of 101 kangaroos.
  (number of insects) variables.
}
\description{
  Variables are \code{sex} (m/f), \code{spec} (giganteus, melanops,
  fuliginosus) and 18 numeric measurements.
}
\source{
  Andrews and Herzberg (1985) Chapter 53.
}
\references{
  Andrews, D. F. and Herzberg, A. M. (1985).
  Data.
  Springer-Verlag, New York.
}
\keyword{datasets}
