\name{lfeval}
\alias{lfeval}
\title{
Extract Locfit Evaluation Structure.
}
\usage{
lfeval(object)
}
\description{
  Extracts the evaluation structure from a \code{"locfit"} object.
  This object has the class \code{"lfeval"}, and has its own set of
  methods for plotting e.t.c.
}
\arguments{
  \item{object}{\code{"locfit"} object}
}

\value{
  \code{"lfeval"} object.
}

\seealso{
  \code{\link{locfit}},
  \code{\link{plot.lfeval}},
  \code{\link{print.lfeval}}
}

\keyword{locfit}
