\name{mcyc}
\alias{mcyc}
\title{Acc(De?)celeration of a Motorcycle Hitting a Wall}
\usage{data(mcyc)}
\format{
  Data frame with time and accel variables.
}
\description{
  Measurements of the acceleration of a motorcycle as it hits a wall.
  Actually, rumored to be a concatenation of several such datasets.
}
\source{
  H\"{a}rdle (1990).
}
\references{
  H\"{a}rdle, W. (1990). Applied Nonparametric Regression.
  Cambridge University Press.
}
\keyword{datasets}
