\name{mine}
\title{Fracture Counts in Coal Mines}
\usage{data(mine)}
\alias{mine}
\format{
A dataframe with the response frac, and predictor variables
extrp, time, seamh and inb.
}
\description{
The number of fractures in the upper seam of coal mines, and four
predictor variables. This dataset can be modeled using Poisson regression.
}
\source{
Myers (1990).
}
\references{
Myers, R. H. (1990). Classical and Modern Regression with Applications
(Second edition). PWS-Kent Publishing, Boston.
}
\keyword{datasets}
