\name{mmsamp}
\alias{mmsamp}
\title{Test dataset for minimax Local Regression}
\usage{data(cltest)}
\format{
Data Frame with x and y variables.
}
\description{
50 observations, as used in Figure 13.1 of Loader (1999).
}
\references{
  Loader, C. (1999). Local Regression and Likelihood. Springer, New York.
}
\keyword{datasets}
