\name{morths}
\alias{morths}
\title{Henderson and Sheppard Mortality Dataset}
\usage{data(morths)}
\format{
  Data frame with age, n and number of deaths.
}
\description{
  Observed mortality for 55 to 99.
}
\source{
  Henderson and Sheppard (1919).
}
\references{
  Henderson, R. and Sheppard, H. N. (1919). Graduation of mortality and
  other tables. Actuarial Society of America, New York.
}
\keyword{datasets}
