\name{penny}
\title{Penny Thickness Dataset}
\usage{data(penny)}
\alias{penny}
\format{
  A dataframe.
}
\description{
  For each year, 1945 to 1989, the thickness of two U.S. pennies
  was recorded.
}
\source{
  Scott (1992).
}
\references{
  Scott (1992). Multivariate Density Estimation.
    Wiley, New York.
}
\keyword{datasets}
