\name{print.gcvplot}
\alias{print.gcvplot}
\title{
Print method for gcvplot objects
}
\usage{
\method{print}{gcvplot}(x, ...)
}
\description{
  Print method for \code{"gcvplot"} objects. Actually, equivalent to
  \code{\link{plot.gcvplot}()}.
  \code{\link{scb}} function.
}
\arguments{
\item{x}{ \code{gcvplot} object.  }
  \item{...}{Arguments passed to and from other methods.}
}
\seealso{
\code{\link{gcvplot}},
\code{\link{plot.gcvplot}}
\code{\link{summary.gcvplot}}
}
\keyword{methods}
