\name{print.lfeval}
\alias{print.lfeval}
\title{
Print the Locfit Evaluation Points.
}
\usage{
print.lfeval(x)
}
\description{
  Prints a matrix of the evaluation points from a \code{"locfit"}
  or \code{"lfeval"} structure.
}
\arguments{
  \item{x}{A \code{"lfeval"} or \code{"locfit"} object}
}

\value{
  Matrix of the fit points.
}

\seealso{
  \code{\link{lfeval}},
  \code{\link{locfit}},
  \code{\link{plot.lfeval}}
}

\keyword{locfit}
