\name{print.locfit}
\alias{print.locfit}
\title{
Print method for "locfit" object.
}
\usage{
print(object, \dots)
}
\description{
Prints a short summary of a \code{"locfit"} object.
}
\arguments{
\item{object}{
\code{"locfit"} object.
}
}
\seealso{\code{\link{locfit}}}
\keyword{locfit}
% Converted by Sd2Rd version 0.2-a5.
