\name{print.preplot.locfit}
\alias{print.preplot.locfit}
\title{
Print method for preplot.locfit objects.
}
\usage{
\method{print}{preplot.locfit}(x, ...)
}
\description{
  Print method for objects created by the
  \code{\link{preplot.locfit}} function.
}
\arguments{
\item{x}{ \code{"preplot.locfit"} object.  }
\item{...}{not used.}
}
\seealso{
\code{\link{preplot.locfit}},
\code{\link{predict.locfit}}
}
\keyword{methods}
