\name{print.summary.locfit}
\alias{print.summary.locfit}
\title{
Print a Locfit summary object.
}
\usage{
\method{print}{summary.locfit}(x, ...)
}
\description{
  Print method for \code{"summary.locfit"} objects.
}
\arguments{
\item{x}{
Object from \code{\link{summary.locfit}()}.
}
\item{...}{not used.}
}
\seealso{
\code{\link{summary.locfit}()}
}
\keyword{methods}
