\name{spencer}
\alias{spencer}
\title{Spencer's Mortality Dataset}
\usage{data(spencer)}
\alias{spencer}
\format{
  Data frame with age and mortality variables.
}
\description{
  Observed mortality rates for ages 20 to 45.
}
\source{
  Spencer (1904).
}
\references{
  Spencer, J. (1904).
  On the graduation of rates of sickness and mortality.
  Journal of the Institute of Actuaries 38, 334-343.
}
\keyword{datasets}
