\name{stamp}
\alias{stamp}
\title{Stamp Thickness Dataset}
\usage{data(stamp)}
\format{
  Data frame with \code{thick} (stamp thickness) and \code{count}
  (number of stamps) variables.
}
\description{
  Thicknesses of 482 postage stamps of the 1872 Hidalgo issue of Mexico.
}
\source{
  Izenman and Sommer (1988).
}
\references{
  Izenman, A. J. and Sommer, C. J. (1988).
  Philatelic mixtures and multimodal densities.
  Journal of the American Statistical Association 73, 602-606.
}
\keyword{datasets}
