\name{store}
\alias{store}
\title{
Save S functions.
}
\usage{
store(data=FALSE, grand=FALSE)
}
\description{
  I've gotta keep track of this mess somehow!
}
\arguments{
  \item{data}{should the data sets be dumped to file?}
  \item{grand}{should the whole shebang be dumped to file?}
}
\keyword{internal}
