\name{summary.preplot.locfit}
\alias{summary.preplot.locfit}
\title{
Summary method for a preplot.locfit object.
}
\usage{
summary(object, \dots)
}
\description{
Prints a short summary of a \code{"preplot.locfit"} object.
}
\arguments{
\item{object}{
\code{"preplot.locfit"} object.
}
}

\value{
The fitted values from a
\code{"preplot.locfit"} object.
}

\keyword{locfit}
