\name{trimod}
\alias{trimod}
\title{Generated sample from a bivariate trimodal normal mixture}
%\usage{data(trimod)}
\format{
  Data frame with 225 observations and variables x0, x1.
}
\description{
  This is a random sample from a mixture of three bivariate standard
  normal components;
  the sample was used for the examples in Loader (1996).
}
\source{
  Randomly generated in S.
}
\references{
  Loader, C. R. (1996).
  Local Likelihood Density Estimation.
  Annals of Statistics 24, 1602-1618.
}
\keyword{datasets}
