\name{formula.locfit}
\alias{formula.locfit}
\title{
Formula from a Locfit object.
}
\usage{
\method{formula}{locfit}(x, ...)
}
\description{Extract the model formula from a locfit object.}
\arguments{
  \item{x}{ \code{locfit} object.}
  \item{...}{Arguments passed to and from other methods.}
}
\value{
Returns the formula from the locfit object.
}
\seealso{
\code{\link{locfit}}
}
\keyword{models}
% Converted by Sd2Rd version 0.2-a5.
