\name{panel.xyplot.lf}
\alias{panel.xyplot.lf}
\title{
Locfit panel function
}
\usage{
panel.xyplot.lf(x, y, subscripts, clo, cup, wh, type, ...)

}
\description{
Panel function used by \code{\link{plot.locfit.3d}} for one dimensional
plots.
}
\seealso{
  \code{\link{plot.locfit.3d}}
}

\keyword{locfit}
