\name{plot.locfit.1d}
\title{
Plot a one dimensional "preplot.locfit" object.
}
\usage{
plot.locfit.1d(x, add, \dots)
}
\arguments{
\item{x}{
One dimensional \code{"preplot.locfit"} object.
}
\item{add}{
If \code{TRUE}, the plot will be added to the existing plot.
}
\item{...}{
Additional arguments to the \code{\link{plot}} function.
}
}
\description{
This function is not usually called directly. It will be called automatically
when plotting one-dimensional \code{"locfit"} or \code{"preplot.locfit"}
objects.
}
\seealso{
\code{\link{locfit}}, \code{\link{plot.locfit}}, \code{\link{preplot.locfit}}
}
\keyword{locfit}
