\name{print.locfit}
\alias{print.locfit}
\title{
Print method for "locfit" object.
}
\usage{
\method{print}{locfit}(x, \dots)
}
\description{
Prints a short summary of a \code{"locfit"} object.
}
\arguments{
\item{x}{\code{locfit} object.}
  \item{...}{Arguments passed to and from other methods.}
}
\seealso{\code{\link{locfit}}}
\keyword{methods}
% Converted by Sd2Rd version 0.2-a5.
