\name{print.locfit}
\alias{print.locfit}
\title{
Print method for "locfit" object.
}
\usage{
\method{print}{locfit}(x, ...)
}
\description{
Prints a short summary of a \code{"locfit"} object.
}
\arguments{
\item{x}{
\code{"locfit"} object.
}
\item{...}{not used.}
}
\seealso{\code{\link{locfit}}}
\keyword{smooth}
% Converted by Sd2Rd version 0.2-a5.
