\name{print.scb}
\alias{print.scb}
\title{
Print method for simultaneous confidence bands
}
\usage{
print.scb(x, ...)
}
\description{
  Print method for simultaneous confidence bands created by the
  \code{\link{scb}} function.
}
\arguments{
\item{x}{ \code{"scb"} object created by \code{\link{scb}()}.  }
}
\seealso{
\code{\link{scb}}
}
\keyword{methods}
