\name{summary.locfit}
\alias{summary.locfit}
\title{
Print method for a locfit object.
}
\usage{
\method{summary}{locfit}(object, ...)
}
\description{
Prints a short summary of a \code{"locfit"} object.
}
\arguments{
\item{object}{
\code{"locfit"} object.
}
\item{...}{currently not used.}
}

\value{
A \code{"summary.locfit"} object, containg a short summary of the
\code{"locfit"} object.
}

\keyword{smooth}
% Converted by Sd2Rd version 0.2-a5.
