\name{summary.preplot.locfit}
\alias{summary.preplot.locfit}
\title{
Summary method for a preplot.locfit object.
}
\usage{
\method{summary}{preplot.locfit}(object, ...)
}
\description{
Prints a short summary of a \code{"preplot.locfit"} object.
}
\arguments{
\item{object}{
\code{"preplot.locfit"} object.
}
\item{...}{currently not used.}
}

\value{
The fitted values from a
\code{"preplot.locfit"} object.
}

\keyword{smooth}
